@@ -15,6 +15,7 @@
         the following property is included to remove the above warning during the build.
         -->
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <java.version>17</java.version>
 
         <!-- Dependency Versions  -->
         <junit.version>3.8.1</junit.version>
